Add left_right::Writer::flush
A Future based flushing call.

To support both a thread based and task::Waker based waking system I've
added the Waker type. This type encapsulates the complexities of the
waker, providing an easy to use API to the users.

To ensure the Writer is always in a valid state the returned Future will
block the current thread if it was not polled to completion.